pyproject.toml
[project]
name = "kokoro-tts-mcp"
version = "0.1.0"
description = "Kokoro Text to Speech (TTS) MCP Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kokoro-onnx>=0.4.5",
    "soundfile>=0.13.1",
    "boto3>=1.37.18",
    "mcp[cli]>=1.26",
    "python-dotenv>=1.0.0",
    "flask==2.2.3",
    "flask-cors==3.0.10",
    "werkzeug==3.1.3",
    "numpy>=1.26.0",
]
